DDL + script to dump it
Will be used in the future to easily sync and track DDL changes
e.g. if we want to drop a column, we just commit the DDL with the
dropped column, this way we track in git the DDL changes and can keep
commits and PR complete when a change impact both the code and the
database.