Delete unused commentsnamecut setting
I noticed I could not save my profile.
It's because this setting is 50 by default, but is set to 25 max on the
account page.
It prevents me to save the account if I don't lower it manually.

11600 users have it set to 25+ vs 52 users who have it set <= 25.

As it does not look like it's being used, I'd rather kill it for good.

After going like, the one who deploys it should run:
```sql
ALTER TABLE `usersettings`
DROP COLUMN `commentsnamecut`;
```